refactor: extract utility command handlers to CommandHandlers.cs
